Add the possibility to whitelist SAML Identity Providers
Some Identity Providers (SAML) returns assertions that are not signed. This may
or may not be due to version mismatch with SAML.

An unsigned SAML response means that we've no idea whether the user is really
authenticated or not.

Up until now we've accepted unsigned assertion, only logging a warning. This
patch changes that behaviour to refuse unsigned assertions - unless the Identity
Provider is whitelisted, in which case we use the old behaviour (log warning and accept).