Make all sql lowercase in JdbcUsers
SQL is by default case-insensitive when using unquoted names. Quoted names must
match exactly.

The problem is that unquoted column names are folded into either lowercase _or_
uppercase depending on database. Postgresql folds into lowercase and Hypersql
folds into uppercase.

The maximum interoperability then occurs when either
  + All column names are quoted at all times.
  + No column names are ever quoted.

The first is very verbose and kind of hard to read (This is Scala syntax):
  """SELECT "NAME" FROM "USERS""""

While the second is easier on the eyes:
  "select name from users"